Separated collisionDetection.cpp into multiple smaller files. Added seperatingAxisTests to Collisions class.
